% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_hillshade}
\alias{create_hillshade}
\title{Create Hillshade from SpatRaster-DEM}
\usage{
create_hillshade(dem, angle = 45, direction = 315)
}
\arguments{
\item{dem}{Digital elevation modell of class SpatRaster}

\item{angle}{Angle of shading}

\item{direction}{Light direction}
}
\description{
Create Hillshade from SpatRaster-DEM
}
