% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_surface.R
\name{create_cost_surface}
\alias{create_cost_surface}
\title{Create Cost Surface}
\usage{
create_cost_surface(
  dem,
  epsg = NULL,
  numberOfNeighbors = 16,
  numberOfDirections = 16,
  slopeGainFactor = FALSE,
  slopeGainStart = 0.1,
  slopeBarrier = FALSE,
  slopeBarrierValue = 0.07,
  barriers = NULL,
  wetlands = NULL,
  wetlandsFactor = 1.78,
  costFunction = "ToblersHikingFunction",
  critical_slope = 10
)
}
\arguments{
\item{dem}{Raster. Digital elevation data. Best as class SpatRaster; sp, sf and raster objects may also work.}

\item{epsg}{EPSG-Code as string (e. g. "EPSG:25832").}

\item{numberOfNeighbors}{Whole number. Number of neighbours to be considered when calculating slope (16 is default).}

\item{numberOfDirections}{Whole number. Number of neighbours for cost surface calculation (4, 8 or 16).}

\item{slopeGainFactor}{TRUE/FALSE. Default FALSE. If TRUE, a quadratic function starting slopeGainStart value will be used}

\item{slopeGainStart}{A number.  Start slope value of the quadratic function in slope m (e.g. 0.1 for 10\% slope)}

\item{slopeBarrier}{TRUE/FALSE. Default FALSE. If TRUE, slope higher than given slopeBarrierValue get NA values (not passable)}

\item{slopeBarrierValue}{Start slope value for barriers (e.g. 0.07 for 7\% slope).}

\item{barriers}{Polygon of class SpatVector. Area of barriers (e.g. a river, lake, political boundaries)}

\item{wetlands}{Polygon of class SpatVector. Area of wetlands/fens.}

\item{wetlandsFactor}{Factor of wetlands-friction (e.g. 0.9 for factor 10); calculated directly on speed}

\item{costFunction}{Cost function. Implemented: "ToblersHikingFunction",  "Irmischer-Clarke's", "Wheeled-Vehicels"}

\item{critical_slope}{Critical Slope, only relevant for Wheeled Vehicles function. Default is 10\%.}
}
\value{
Cost surface
}
\description{
Create Cost Surface
}
