% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_surface.R
\name{create_cost_surface}
\alias{create_cost_surface}
\title{Create Cost Surface}
\usage{
create_cost_surface(
  dem,
  epsg = NULL,
  numberOfNeighbors = 16,
  numberOfDirections = 16,
  slopeGainFactor = FALSE,
  slopeGainStart = 0.1,
  slopeBarrier = FALSE,
  slopeBarrierValue = 0.07,
  barriers = NULL,
  wetlands = NULL,
  wetlandsFactor = 1.78,
  costFunction = "ToblersHikingFunction",
  critical_slope = 10
)
}
\arguments{
\item{dem}{Raster. Digital elevation data. Best as class SpatRaster; sp, sf and raster objects may also work.}

\item{epsg}{EPSG-Code as string (e.g. "EPSG:25832"). Currently not used internally.}

\item{numberOfNeighbors}{Whole number. Number of neighbours to be considered when calculating slope (16 is default).}

\item{numberOfDirections}{Whole number. Number of neighbours for cost surface calculation (4, 8 or 16).}

\item{slopeGainFactor}{Logical. Default FALSE. If TRUE, a quadratic function starting at \code{slopeGainStart} will be used.}

\item{slopeGainStart}{Numeric. Start slope value of the quadratic function in slope m (e.g. 0.1 for 10\% slope).}

\item{slopeBarrier}{Logical. Default FALSE. If TRUE, slopes higher than \code{slopeBarrierValue} get NA values (not passable).}

\item{slopeBarrierValue}{Numeric. Start slope value for barriers (e.g. 0.07 for 7\% slope).}

\item{barriers}{Polygon of class SpatVector. Area of barriers (e.g. rivers, lakes, political boundaries).}

\item{wetlands}{Polygon of class SpatVector. Area of wetlands/fens.}

\item{wetlandsFactor}{Numeric. Factor of wetlands friction (e.g. 1.78); applied directly on speed (1 / wetlandsFactor).}

\item{costFunction}{Character. Cost function. Implemented: "ToblersHikingFunction", "Irmischer-Clarke's", "Wheeled-Vehicels".}

\item{critical_slope}{Numeric. Critical slope in percent (only relevant for "Wheeled-Vehicels"). Default is 10.}
}
\value{
A \code{TransitionLayer} object representing the cost surface.
}
\description{
Creates a cost surface based on a digital elevation model (DEM) using
different cost functions and optional barrier or wetland factors.
}
