% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsc_territory.R
\name{lcsc_territory}
\alias{lcsc_territory}
\title{LCSC Territory - Potential settlement area}
\usage{
lcsc_territory(
  dem,
  sites,
  movement_time,
  max_speed,
  slopeBarrier = FALSE,
  slopeBarrierValue = NULL,
  write_polygon = FALSE,
  write_raster = FALSE,
  name_index = NULL,
  wd = NULL,
  epsg = "EPSG:25832",
  costFunction = "ToblersHikingFunction"
)
}
\arguments{
\item{dem}{Raster. Digital Elevation Data.}

\item{sites}{Sites to analyse as points.}

\item{movement_time}{Time-Frame in minutes}

\item{max_speed}{Maximum possible speed of the cost function (e.g. 6 km/h for ToblersHickingFunction. Optional, but good for speeding up the cost_surface calculation)}

\item{slopeBarrier}{TRUE/FALSE. Default FALSE. If TRUE, slope higher than given slopeBarrierValue get a slope value of NA}

\item{slopeBarrierValue}{Start slope value for barriers.}

\item{write_polygon}{Write polygon of LCSC territory per site to local directory (specified with wd = )}

\item{write_raster}{Write accumulate cost of LCSC territory per site to local directory (specified with wd = )}

\item{name_index}{Index of a column with the names of the sites. If not specified, a name is automatically generated according to the pattern site_ \link{consecutive number}.}

\item{wd}{Directory in which the polygons and any raster data can be saved. Optional.}

\item{costFunction}{Cost-Function. Possible: "ToblersHikingFunction"}
}
\value{
Polygon-Shapefile. Optional: Raster TIFF
}
\description{
The function generates theoretical territories or catchment areas based on a cost surface for one or more locations.
The catchment areas are calculated on the basis of a certain period of time, which, starting from a central location, the movement can take place in all directions depending on the costs.
}
