% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_tpla.R
\name{perform_tpla}
\alias{perform_tpla}
\title{Total Passability Landscape Analysis (TPLA)}
\usage{
perform_tpla(
  cost_surface,
  center_point,
  radius_tpla,
  number_of_points,
  sigma_density_calc,
  project_crs = NULL,
  keep_lines = FALSE
)
}
\arguments{
\item{cost_surface}{Transition Object. Cost surface (Class: Transition, calculated with the gdistance package)}

\item{center_point}{Center point of the study area. Class of SpatVector (terra).}

\item{radius_tpla}{Integer. Radius for the circle of starting points (in meters), in wich tpla will take place.}

\item{number_of_points}{Integer. Number of starting points on the circle.}

\item{sigma_density_calc}{Number. Standard deviation for the kernel density estimation.}

\item{keep_lines}{TRUE or FALSE. Default is FALSE. If TRUE, the cost-optimal paths will be included in the result object.}
}
\value{
S4-Class or Raster. If keep_lines = TRUE, an S4 object containing the result raster of the kernel density estimation and the cost-optimal paths is returned.
}
\description{
Placeholder: Description to follow (suggestions for research-area, number of points, Spatial-classes etc.)
}
\examples{
# fun_TPLA(cost_surface, center_point, radius_tpla, number_of_points, sigma_density_calc, keep_lines=FALSE)
}
