% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_tpla_line_based.R
\name{perform_tpla_line_based}
\alias{perform_tpla_line_based}
\title{Total Passability Landscape Analysis (TPLA) line based}
\usage{
perform_tpla_line_based(
  cost_surface,
  first_line,
  second_line,
  number_of_points,
  sigma_density_calc,
  project_crs = NULL,
  keep_lines = FALSE
)
}
\arguments{
\item{cost_surface}{Transition Object. Cost surface (Class: Transition, calculated with the gdistance package)}

\item{first_line}{Spatial Line. Class of LINESTRING (sf).}

\item{second_line}{Spatial Line. Class of LINESTRING (sf).}

\item{number_of_points}{Integer. Number of starting points on the circle.}

\item{sigma_density_calc}{Number. Standard deviation for the kernel density estimation.}

\item{keep_lines}{TRUE or FALSE. Default is FALSE. If TRUE, the cost-optimal paths will be included in the result object.}
}
\value{
List or Raster. If keep_lines = TRUE, an S4 object containing the result raster of the kernel density estimation and the cost-optimal paths is returned.
}
\description{
TPLA stands for Total Passability Landscape Analysis. Using least-cost paths and a density calculation,
it calculates regions with high routes or movement potential between two spatial lines.
}
\details{
Placeholder: Add more information (suggestions for research-area, number of points, Spatial-classes etc.)
}
\examples{
# fun_TPLA(cost_surface,first_line, second_line, number_of_points, sigma_density_calc, keep_lines=FALSE)
}
