% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbr_network.R
\name{sbr_network}
\alias{sbr_network}
\title{Site-based (theoretical) route network (sbr_network)}
\usage{
sbr_network(sites, lines, conductance, steps_points, max_speed)
}
\arguments{
\item{sites}{SpatialPointsDataFrame or compatible object. The site locations.}

\item{lines}{SpatialLinesDataFrame. The base lines to start from (known route).}

\item{conductance}{Transition Layer. Cost surface (gidstance).}

\item{steps_points}{Integer. The number of steps (normaly in Meter) for sampling points along the lines.}

\item{max_speed}{Numeric. The maximum speed (optional, can be refined based on use case).}
}
\value{
A SpatialLinesDataFrame object with the calculated routes.
}
\description{
Generates a theoretical route network based on site locations, a given route, based on conductance values.
}
